move mat3inv to matrices.[ch]
Now that mul3inv is no longer called in colorspaces.c, it makes sense
to move it next to mat3SSEinv.
