+++ src/common/matrices.c
@@ -1,6 +1,6 @@
 /*
     This file is part of darktable,
-    Copyright (C) 2011-2020 darktable developers.
+    Copyright (C) 2011-2023 darktable developers.
 
     darktable is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -49,6 +49,47 @@ int mat3SSEinv(dt_colormatrix_t dst, const dt_colormatrix_t src)
 #undef B
   return 0;
 }
+
+#define generate_mat3inv_body(c_type, A, B)                                                                  \
+  int mat3inv_##c_type(c_type *const dst, const c_type *const src)                                           \
+  {                                                                                                          \
+                                                                                                             \
+    const c_type det = A(1, 1) * (A(3, 3) * A(2, 2) - A(3, 2) * A(2, 3))                                     \
+                       - A(2, 1) * (A(3, 3) * A(1, 2) - A(3, 2) * A(1, 3))                                   \
+                       + A(3, 1) * (A(2, 3) * A(1, 2) - A(2, 2) * A(1, 3));                                  \
+                                                                                                             \
+    const c_type epsilon = 1e-7f;                                                                            \
+    if(fabs(det) < epsilon) return 1;                                                                        \
+                                                                                                             \
+    const c_type invDet = 1.0 / det;                                                                         \
+                                                                                                             \
+    B(1, 1) = invDet * (A(3, 3) * A(2, 2) - A(3, 2) * A(2, 3));                                              \
+    B(1, 2) = -invDet * (A(3, 3) * A(1, 2) - A(3, 2) * A(1, 3));                                             \
+    B(1, 3) = invDet * (A(2, 3) * A(1, 2) - A(2, 2) * A(1, 3));                                              \
+                                                                                                             \
+    B(2, 1) = -invDet * (A(3, 3) * A(2, 1) - A(3, 1) * A(2, 3));                                             \
+    B(2, 2) = invDet * (A(3, 3) * A(1, 1) - A(3, 1) * A(1, 3));                                              \
+    B(2, 3) = -invDet * (A(2, 3) * A(1, 1) - A(2, 1) * A(1, 3));                                             \
+                                                                                                             \
+    B(3, 1) = invDet * (A(3, 2) * A(2, 1) - A(3, 1) * A(2, 2));                                              \
+    B(3, 2) = -invDet * (A(3, 2) * A(1, 1) - A(3, 1) * A(1, 2));                                             \
+    B(3, 3) = invDet * (A(2, 2) * A(1, 1) - A(2, 1) * A(1, 2));                                              \
+    return 0;                                                                                                \
+  }
+
+#define A(y, x) src[(y - 1) * 3 + (x - 1)]
+#define B(y, x) dst[(y - 1) * 3 + (x - 1)]
+/** inverts the given 3x3 matrix */
+generate_mat3inv_body(float, A, B)
+
+int mat3inv(float *const dst, const float *const src)
+{
+  return mat3inv_float(dst, src);
+}
+
+#undef B
+#undef A
+#undef generate_mat3inv_body
 
 // clang-format off
 // modelines: These editor modelines have been set for all relevant files by tools/update_modelines.py
